Refactor pool page specific components
