+++ src/components/pages/pool/PoolActionsCard/PoolActionsCard.vue
@@ -44,7 +44,7 @@
 import { defineComponent, ref } from 'vue';
 import InvestForm from '@/components/forms/pool_actions/InvestForm.vue';
 import WithdrawForm from '@/components/forms/pool_actions/WithdrawForm.vue';
-import SuccessOverlay from '../shared/SuccessOverlay.vue';
+import SuccessOverlay from '@/components/cards/SuccessOverlay.vue';
 import { useI18n } from 'vue-i18n';
 import TradeSettingsPopover, {
   TradeSettingsContext
