+++ src/components/pages/pool/PoolActivitiesCard/types.ts
