+++ src/components/pages/pool/index.ts
@@ -1,0 +1,5 @@
+export { default as PoolChart } from './PoolChart.vue';
+export { default as PoolStatCards } from './PoolStatCards.vue';
+export { default as PoolBalancesCard } from './PoolBalancesCard.vue';
+export { default as PoolActionsCard } from './PoolActionsCard/PoolActionsCard.vue';
+export { default as PoolActivitiesCard } from './PoolActivitiesCard/PoolActivitiesCard.vue';
